Debug

+++ Tasks.py
@@ -4,7 +4,7 @@
 class Tasks(object):
     def __init__(self,
                  name: int,
-                 work_load: int,  # (内存)
+                 work_load: float,  # (内存)
                  limit1: list,  # 全部用序号表示
                  limit2: list,
                  limit3: list,
